requirements.txt
argon2-cffi==23.1.0
cryptography==42.0.5
pyotp==2.9.0
qrcode==7.4.2
